Functions for calculating mean value are placed in func_mean.py
