+++ ipt_connect/IPTdev/tests.py
@@ -1,5 +1,5 @@
 from django.test import TestCase
-from models import *
+from func_mean import *
 
 # Create your tests here.
